Adding tarfile member sanitization to extractall() (#4)
This commit fixes the case where the codebase is changed such that the [hashes used to validate the input archives](https://github.com/mdegans/tegrity/blob/69e85ceeb6ea18414fda82c4c376ba89a41fce93/tegrity/settings.py#L20) somehow aren't checked, and I don't honestly remember if there's a `--option` somewhere to do that (I don't think so), so it's a good idea to fix this.

This isn't believed to be an *immediate* security issue for `tegrity` because other than the rootfs, which the developer would create themselves, all archives are checked against a SHA_512, and if the developer created a compromised rootfs, that developer would already be compromised. It's possible that somebody could be tricked into downloading, say, an emulator image or something, and paste in a bad SHA_512, then sure, the host could be compromised, but I can't think of everything and it's better to patch this anyway.

That, being said, looking at that code, the **kernel is probably ancient** and loaded with vulnerabilities and I should probably flag this project as abandoned. I might resume development on `dev` at some point if there was some money in it, but otherwise, I'll put a `do not use for now`. I made the project mostly because Nvidia didn't support a particular camera at the time officially, and now they do, so I don't need it anymore personally.